// @types/generated/contentful.d.ts
export interface IAuthor extends Entry<IAuthorFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "author";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface IBlogPostFields {
  /** Title */
  title: string;

  /** Slug */
  slug: string;

  /** Publish Date */
  publishDate: string;

  /** Body */
  body: Document;

  /** Hero Image */
  heroImage?: Asset | undefined;

  /** Author */
  author: IAuthor;

  /** Excerpt */
  excerpt: string;

  /** Featured */
  featured?: boolean | undefined;
}

/** Content type for creating a blog post */

export interface IBlogPost extends Entry<IBlogPostFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "blogPost";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}
// ...
export interface ICardFields {
  /** Internal Name */
  internalName: string;

  /** URL */
  url?: string | undefined;

  /** Open url in new window? */
  isExternal?: boolean | undefined;

  /** Card Header */
  cardHeader?: string | undefined;

  /** Card Body */
  cardBody?: Document | undefined;

  /** Card Footer */
  cardFooter?: Document | undefined;
}

export interface ICard extends Entry<ICardFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "card";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface ICardGridSectionFields {
  /** Title */
  title?: string | undefined;

  /** Subtitle */
  subtitle?: string | undefined;

  /** Cards */
  cards: (ICard | IBlogPost | IEvent | IPressArticle)[];
}

export interface ICardGridSection extends Entry<ICardGridSectionFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "cardGridSection";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}
// ...
export interface IEventFields {
  /** Title */
  title: string;

  /** Image */
  image?: Asset | undefined;

  /** Date */
  date: string;

  /** Link */
  link: string;

  /** Location */
  location?: string | undefined;
}

export interface IEvent extends Entry<IEventFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "event";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}
// ...
export interface IPressArticleFields {
  /** Title */
  title: string;

  /** Featured */
  featured?: boolean | undefined;

  /** Media Type */
  mediaType?: "Article" | "Podcast" | "Video" | "Keynote" | "Event" | undefined;

  /** Image */
  image?: Asset | undefined;

  /** Description */
  description?: string | undefined;

  /** Publication */
  publication?: string | undefined;

  /** Date */
  date: string;

  /** Link */
  link: string;
}

/** A press article */

export interface IPressArticle extends Entry<IPressArticleFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "pressArticle";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}